refactor: organize packages/core into logical folder structure
- Move core LaunchQL class to core/ folder
- Move module utilities to modules/ folder
- Move resolution logic to resolution/ folder
- Move packaging utilities to packaging/ folder
- Move workspace utilities to workspace/ folder
- Move extension management to extensions/ folder
- Move export functionality to export/ folder
- Update all import paths to reflect new structure
- Maintain backward compatibility through index.ts exports

This reorganization improves code organization and maintainability
while preserving all existing functionality.
