+++ packages/core/src/export/export-meta.ts
