+++ packages/core/src/index.ts
@@ -1,17 +1,17 @@
-export * from './class/launchql';
-export * from './deps';
-export * from './export-meta';
-export * from './export-migrations';
-export * from './extensions';
-export * from './modules';
-export * from './package';
-export * from './paths';
-export * from './resolve';
+export * from './core/class/launchql';
+export * from './resolution/deps';
+export * from './export/export-meta';
+export * from './export/export-migrations';
+export * from './extensions/extensions';
+export * from './modules/modules';
+export * from './packaging/package';
+export * from './workspace/paths';
+export * from './resolution/resolve';
 export * from './projects/deploy';
 export * from './projects/revert';
 export * from './projects/verify';
-export * from './transform';
-export * from './utils';
+export * from './packaging/transform';
+export * from './workspace/utils';
 
 // New exports for migration API
 export * from './migrate/migration';
