+++ packages/core/src/modules/modules.ts
@@ -1,6 +1,6 @@
 import { sync as glob } from 'glob';
 import { basename } from 'path';
-import { getLatestChange, parseControlFile, Module } from './files';
+import { getLatestChange, parseControlFile, Module } from '../files';
 
 export type ModuleMap = Record<string, Module>;
