+++ packages/core/src/projects/deploy.ts
@@ -5,8 +5,8 @@ import { PgConfig } from 'pg-env';
 import { Logger } from '@launchql/logger';
 import { getPgPool } from 'pg-cache';
 import { deployModule } from '../modules/deploy';
-import { LaunchQLProject } from '../class/launchql';
-import { packageModule } from '../package';
+import { LaunchQLProject } from '../core/class/launchql';
+import { packageModule } from '../packaging/package';
 import { runSqitch } from '../utils/sqitch-wrapper';
 
 interface Extensions {
