+++ packages/core/src/resolution/resolve.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from 'fs';
-import { getChanges, getExtensionName } from './files';
-import { parsePlanFile } from './files/plan/parser';
+import { getChanges, getExtensionName } from '../files';
+import { parsePlanFile } from '../files/plan/parser';
 import { resolveDependencies } from './deps';
 
 /**
