+++ packages/core/src/workspace/paths.ts
